split cde to standalone package.
